+++ src/ui/about.rs
@@ -104,6 +104,12 @@ impl SimpleComponent for AboutDialog {
 
             set_release_notes_version: &APP_VERSION,
             set_release_notes: &[
+                "<p>Added</p>",
+
+                "<ul>",
+                    "<li>Added support to the new wishes url cache location</li>",
+                "</ul>",
+
                 "<p>Fixed</p>",
 
                 "<ul>",